requirements.txt
pandas==2.0.2
streamlit==1.24.0
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
wget==3.2
streamlit==1.24.0